facade_service/requirements.txt
pika==1.2.0
uuid==1.30
python-consul==1.1.0
requests==2.25.1
flask==2.0.1